refactor : dto 구조 분리.
